bot/Telegram/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
accelerate
torch
safetensors
requests
pydantic
pillow
pdfplumber
pytesseract
trafilatura
beautifulsoup4
regex
telebot
tiktoken
feedparser
sentence-transformers
numpy
